refactor(scalars): components location
