+++ packages/design-system/src/scalars/components/fragments/form-label/index.ts
