+++ packages/design-system/src/scalars/components/fragments/input/input.stories.tsx
