+++ packages/design-system/src/scalars/components/fragments/tooltip/index.ts
