+++ packages/design-system/src/scalars/components/index.ts
@@ -1,10 +1,6 @@
 export * from "./checkbox";
-export * from "./form-group";
-export * from "./form-label";
-export * from "./form-message";
 export * from "./scalar-demo";
 export * from "./toggle";
 export * from "./radio-group";
 export * from "./toggle/toggle";
-export * from "./input";
 export * from "./string";
